Conversion of integer datatype to 64-bit
Changed implementation of APIs to use 64-bit integer for all dimensional variables.
User facing APIs retain standard signature for compatibility. Only internal implementation changed
by creating compatibility layer.
Only LAPACK interfaces and FLAME interfaces are changed whereas LAPACKE interface remains the same.
All test suites accordingly changed (Netlib, legacy and main test).

+++ src/src_cpp/libflame.hh
@@ -876,11 +876,11 @@ inline void bdsqr(char* uplo, integer* n, integer* ncvt, integer* nru, integer* 
 }
 
 // --- Bidiagonal divide-and-conquor algorithm ---
-inline void bdsdc(char* uplo, char* compq, integer* n, float*  d, float*  e, float*  u, integer* ldu, float*  vt, integer* ldvt, float*  q, float*  iq, float* work, integer* iwork, integer* info)
+inline void bdsdc(char* uplo, char* compq, integer* n, float*  d, float*  e, float*  u, integer* ldu, float*  vt, integer* ldvt, float*  q, integer*  iq, float* work, integer* iwork, integer* info)
 {
   sbdsdc_(uplo, compq, n, d, e, u, ldu, vt, ldvt, q, iq, work, iwork, info);
 }
-inline void bdsdc(char* uplo, char* compq, integer* n, double* d, double* e, double* u, integer* ldu, double* vt, integer* ldvt, double* q, double* iq, double* work, integer* iwork, integer* info)
+inline void bdsdc(char* uplo, char* compq, integer* n, double* d, double* e, double* u, integer* ldu, double* vt, integer* ldvt, double* q, integer* iq, double* work, integer* iwork, integer* info)
 {
   dbdsdc_(uplo, compq, n, d, e, u, ldu, vt, ldvt, q, iq, work, iwork, info);
 }
